requirements-dev.txt
asgiref==3.3.1
braintree==3.59.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-storages==1.9.1
flake8===3.8.4
idna==2.10
isort==5.7.0
Pillow==7.0.0
pytz==2021.1
requests==2.25.1
selenium==3.141.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
sentry-sdk==0.16.5
